Add interface match requirement for service contracts
An operator contract can only be activated on a service contract
if the operator contract's (claimed) implemented interface
matches the service contract.
This prevents cross-activation of operator contracts with conflicting interfaces
on service contracts for a different service,
and ensures that a compromised Operator Contract Upgrader
can cause very little trouble on its own.