Exit from relay request command on error, keep requests separate
Two changes to the relay request CLI command:

1. When there is an error during relay request submission we now exit
instead of hanging forever

2. In the rare case when relay entry submission happens before new relay
request in the same block, we don't want to fulfill the promise with the
value of the generated entry from that block. That's why we need to wait
for relay request created event before we install on relay entry submitted
handler.