IAM user update backlog and new addtions
Markus is no longer with us so we need to remove him from the owner mix.
In addition to the removal we add two new IAM groups, one for role
editor and one for role storage.objectviewer.  The members outlined in
each are part of existing permissions I was lazy with and added via the
UI.