Create separate broadcast channel for DKG execution
BroadcastChannel libp2p implementation synchronizes on mutex when
publishing message to the pubsub and when firing handlers. When we had
one channel for the given list of stakers used for all group operations,
it could happen that it was used at the same time for DKG and relay entry
signing if the same list of stakers was selected to a new group.
As a result, some messages could not reach the destination because mutex
synchronization was slowing down message exchange and the given phase
could end before all messages were send and delivered.

This change separates DKG from relay entry signing. We create a
temporary channel, just for DKG. The name of the final channel for a
group is a group public key in a compressed form.