Move eth-networks into infrastructure
This will make more sense with the next change that's coming.  It dind't
sit entirely well with me that eth-networks was on the keep-core root
dir.  It's clearer to me why now, what goes into this folder will
ultimately be determined by both eth-network and environment, e.g.
keep-dev/test.  That said this should sit in infrastructure along side
the other environmental resources.