RetransmissionOptions replaced with Context
We will fire up a Context for each state machine state. The Context will
be active for the entire duration of the given state. Global network
Ticker will provide ticks and on each tick we will do a retransmission
until the context gets cancelled.

This proposition seems more elegant than the previous idea with
RetransmissionOptions.

For the time being, passing context to Send is optional. This may change
in the future.