Allow to trigger cache sweeping
Previously, cache sweeping was always done automatically when a new
entry was added to the cache. In this commit we exposed sweep function
that can be called to clean up the cache. We will use this function in
one of the next commits to sweep the cache after retransmissions for the
given phase are done.