Context scoped to the lifetime of state
Here we introduce a context scoped to the lifetime of a state and we
pass this context to state's Initiate function. Context is used to
register state-bound broadcast channel receiver. We will also use this
context later for request-scoped broadcast channel Send invocation.

For any protocol, each state exchanging messages consist of three
phases: delay phase, initialization phase, and active phase. During the
delay phase, we do nothing - this is an additional time given for every
participant to enter the state. During initialization, we send all the
messages. During the active phase, we wait for a certain time for
messages from all other participants to arrive.

All transitions are synchronized on new blocks appearing on-chain.

Since libp2p pubsub does not guarantee delivery, we retransmit messages
for a certain number of times with the configured interval to increase
the message delivery rate. Retransmission options are configurable in
the toml file.

Retransmissions are time-synchronized and state machine transitions are
block-synchronized. That's the first problem.

Another problem is with the delay phase - if we assume two or more
clients can receive delayed information about new blocks, there is no
guarantee one or two blocks for the delay phase are enough to let them
sync, especially with very short block times. What's more, if we assume
retransmissions will help to alleviate the pain with imperfect
synchronizations, the question about the need for the existence of the
delay phase arises.

There is a remedy.

We need to bind together state transitions and retransmissions of
network messages. We need to assume broadcast channel works in a "beacon
mode" retransmitting message with certain intervals for the entire
lifetime of the given state.

To make this happen, we need to make Send and Recv context-aware.

Send will retransmit a message for the entire lifetime of the context.
Recv will receive message filtering out duplicates for the entire
lifetime of the context.