Eliminate possibility of receiving message after cancelling the context
Go language specification says that for select statement if one or more
of the communications can proceed, a single one that can proceed is
chosen via a uniform pseudo-random selection.

If we look at the previous code of goroutine inside of the Recv function,
we can notice that calling handler and removing is are both possible and
depends on the result of the mentioned pseudo-random selection in case
when the context has been cancelled.

In Recv function documentation we say that "when the context is done,
handler is automatically unregistered and receives no more messages".

The change in this PR changes the channel logic to honor this statement.