retransmission.WithRetransmissionSupport
WithRetransmissionSupport takes the standard network message handler and
enhances it with functionality allowing to handle retransmissions. The
returned handler filters out retransmissions and calls the delegate
handler only if the received message is not a retransmission or if it is
a retransmission but it has not been seen by the original handler yet.