Retire internal Keep Network in keep-test
Some weeks ago we've deployed a second Keep Network to keep-test.  This
second network is deployed to Ropsten.  As part of that we've issued
updated bundles to our network participants, that includes new
configs / connections to Ropsten.  Now that this new deployment has been
running for a reasonable amount of time we're going to retire the Keep
Network deployment that connects to our internal ETH network in
keep-test.  This will simplify our setup and reduce operational
overhead.