Replace operatorProfiles with operatorAddresses
Staking delegations were recently updated to be calculated via the
operator and contract owner public addresses, instead of signatures.
With this change we no longer need operator address + private key and
contract owner + private key to facilitate staking.  We've removed the
"profile" key from operatorProfiles and updated the values to include
only addresses.  Changes to the staking function and the forEach reflect
the change in operatorProfiles structure.