WatchBlocks function for BlockCounter
WatchBlocks returns a channel that will emit new block numbers as they
are mined. When the context provided as the parameter ends, new blocks
are no longer pushed to the channel and the channel is closed. If there
is no reader for the channel or reader is too slow, block updates can be
dropped.

Here we provide two implementations: ethereum and local.

WatchBlocks will be used to trigger network message retransmissions
periodically.