Fixed blockcounter's WatchBlock test
Since the channel returned by ethereum block counter's WatchBlock
function is no longer buffered it may happen we write to the
subscription channel before watcher goroutines are registered. In such
case, by design, message is dropped by the channel returned from
WatchBlock. It is enough to wait for some time to give the test
receivers a chance to set up.