Peer outbound queue size is 128, use default number of validate workers
We don't want to use more validate workers for synchronous validation
(signature) than the number of CPUs. This can be only slower because of
the context switching. Instead, we'll increase the size of validate
queue (the next commit).

Also, we are increasing outbound queue size from 64 to 128 in case two
retransmission ticks are very close to another and we need to accommodate
2*64 messages in that queue.