First pass KEEP faucet as Google Function
Here we provide a KEEP staking function expressed as a JS Google Cloud
Function.  The purpose here is to provide a minimally functional facuet
for issuing and staking KEEP while keeping operational overhead in mind.
Minimally functional here being that for a given address we can issue
and stake KEEP on demand from a public interface.