Switch back to libp2p/go-libp2p-pubsub
We had to switch to our fork because of the PR we submitted to libp2p
allowing to configure size of validation queue with
WithValidateQueueSize function.

https://github.com/libp2p/go-libp2p-pubsub/pull/255/

This PR is merged so we can go back to the libp2p/go-libp2p-pubsub.
Since there was no official release from the time our PR was merged, we
are attached to a commit in go-libp2p-pubsub master.