Differentiate between active and eligible stake
Active stake is a stake that passed the initialization period and lets
the operator to join the network and potentially finish any outstanding
work.

Eligible stake is an active stake that is not currently undelegating. It
lets the operator to join the network and makes the operator eligible
for work selection.