Remove active groups threshold
The reason for having the active groups threshold was to not go down to
zero groups and render the beacon unresponsive.

However,
This could still happen if groups were not responding to requests.
This could not happen if beacon was serving enough requests.

Also,
This was easy to repair - just deploy another operator contract and
genesis.

And more importantly,
This was conflicting with unstaking.

Operator can undelegate their stake, wait for the undelegation period
and that's all! Operator is gone from the staking contract but it is
still a group member!

The solution is to make group active time equal to stake undelegation
period and always expire old groups.

Here we take care of the latter.