Callback reimbursement cannot be higher than the gas limit
In case of heavily underpriced transactions, EVM may wrap the 'call' with
additional opcodes. In this case, gas spent > gas limit and we return
just the gas limit.
The worst case scenario cost has to be included in entry verification fee.
Based on the performed tests, it's never more than ~1200gas now.