Remove ETH account creation
Many moons ago we had a notion that we may want or need to quickly scale
out keep-clients for testing.  It turns out this never happened and we
really don't need to make sure on the deployment level, that we can
generate an ETH account.  So, we can go ahead and get rid of account
creation in the provisioning script.  This will be coupled with a few
other changes to make sure an assigned account is funded and staked.