Client interfaces for random beacon and token staking
Applications integrating with Keep Random Beacon need to have an
interface to work with. Since the service contract is hidden behind a
proxy, the most elegant solution is to expose a public, minimal
interface the application can use against the proxy address.

We do a similar thing for the token staking contract except that the
interface of token staking contract is intended to be used by other Keep
network systems rather than external applications.