Use HDWalletProvider for migrations
I'm trying to standardize methods here.  For Ropsten migrations we need
to use the HDWalletProvider, so I'm extending that to our other Kube
migrations.  Along with that we want to keep the Contract Owner keys out
of source control.  We'll use a Circle Context for the key and pass that
in via an env var to the truffle config.