Migrate keep-client-0 to new provisioning
We've updated our InitContainer, and as a result we need to update our
keep-client StatefulSet configs.  There were a couple reasons behind
this re-framing, but the gist is that we wanted to get app configuration
data out of opaque parts of the config pipeline and put somewhere where
the operator of an environment can get to the quickly and see what's
set.