Deleted ITokenStaking
If we make sortition pools depend on keep-core, we'll get a circular
dependency when we'll try to use sortition-pool from keep-core.

We decided it is better to let sortition-pool define expectations about
bonding and staking interface rather than importing them from keep-core.

For this reason, we can delete ITokenStaking as no external customer
application should be interested in this functionality.