Quit with all the streaming business
Thanks to some help from Kristen and Antonio we landed on the previous
method just chunking everything into a blob (via concat) for processing
anyway, so why deal with the callback I wasn't handling anyway.  Here
we switch to readFileSync and writeFileSync, which have the build in
handling to complete before process.exit() is called.