Simplify eth url construction
There are a couple things in play here. Not all eth hosts are
configured in the format hostname:port.  Hosted solutions such as
Infura we just pass url provided by the host.  This complicates
things when we're setting config values with a host and port
variable.  Our connections to internal hosts need a host:port
configured.  We could probably address this internally by flipping
our ETH node Service configs over to an 80 -> backend port mapping
but that's going to be a pretty significant rejiggering.  Instead we're
combining the host:port variables into a single URL var for the RPC and
WS connections.  This will allow us to include the port in the URL var
when needed, and leave it out when not.  Now it's up to the operator to
know when they need to do this...but I guess they should know anyway
:shrug: