Remove authorization check
The function I need to call in the authorization check isn't available
on the tag that we're deploying, so I'm removing the check.  Running the
authorization against a operator contract / operator pair that's already
authorized doesn't appear to break things so we're ok from a
provisioning perspective without the check.  Although it does add extra
time to the InitContainer run.