Script adjustments for config re-org
Each staker tester has their own account passphrase now, as a result we
had to move their passphrase under their respective /participant
directory.  This needed to be accounted for in the bundling scripts.
The rest of the scripts were simplified due to the recent overall config
changes.  tl;dr here is that the only scenario we need to account for in
this scripts at the moment is keep-test / ropsten.  The rest are handled
by automated config now.