Do not delete operator's information after cancelling/recovering stake
With `delete operators[_operator]` as before, if one decides to cancel,
they can't withdraw unbounded ETH from bonding contract because beneficiary
info is gone.

To make things even worse, another party can - in the meantime - delegate
to the same operator and withdraw ETH from bonding to their beneficiary.
That another party would be probably a front-runner bot.

Another problematic scenario is when one first recovers the stake and then
tries to withdraw beacon rewards. It is not possible because the beneficiary
info is gone.

With this change, we dont delete `operators[_operator]`, thus we don't
allow to reuse the same operator address even if staking has been cancelled
or stake has been recovered. Delegation (operator) information stays in
the contract forever so rewards or bonding can be withdrawn at any
moment without a risk.