Improved test coverage of TokenGrant contract
This commit is huge but I could not split it into smaller ones without
first removing all tests. The only test we had for
stake/cancel/undelegate/recover path was long and not covering all
paths. Also, it did not work in separation - some other test was setting
up data for it.