Group test files with directories and shorten their names
Long, redundant names for the .js test files are difficult to work with.
Moving the files to directories based on common prefixes
(KeepRandomBeaconOperator and KeepRandomBeaconService)
and dropping the redundant parts means that the relevant information
(what particular functionality is tested in the file)
is not obscured by prefixes.

For example, on Github only the prefixes of the file names are rendered,
making looking for a particular file extremely frustrating.
With this change, Github will display what the test file is actually doing.