demo.js renamed to delegate-tokens, removed grant parts
We used demo.js for a long time as a way to initialize local development
environments. demo.js was designed to set up KEEP token dashboard dApp a
long time ago. For local developmnent, we don't need grants. For a quick
dApp demo setup, we do.

Here we strip demo.js to only the required token delegations part and
rename it to delegate-tokens.js. We will develop a separate script for
staking dApp setup and put this code in a separate file.