Increase group selection gas estimate
Current group selection gas estimate doesn't
include a corner case when group selection
is triggered after previous DKG failed
and some related data (previous tickets)
must be cleaned up in the same transaction.
This cause a gas peak which is returned
after cleanup is done but the gas estimate
must include it in order to avoid `out of gas`
error. Increasing the value of
`groupSelectionGasEstimate` to `200k` should help.