Remove faucet guide
The PR that implements the faucet was never reviewed/merged.  It's now
out of date and needs work before it can be used with the latest version
of contract interfaces.  Removing the guide for now to prevent
confusion.  It can be added back if we decide to implement the faucet.