Moved implementation initialization to the second upgrade step
Previously, we initialized implementation contract in the first step of
the upgrade and that was giving us some possibility to interact with the
new contract before it should actually be used. In the worst case, this
could work as an upgrade backdoor on our side allowing us to upgrade the
system faster, without keeping the 1 day delay promise.

Here, we move the initialization call to complete upgrade function.

We use a mapping to store implementation initialization data as it is
allocating storage based on keccak256 hash of the key and entry so it
lowers the possibility of clashing with implementation fields.

Complete upgrade function can be now executed only by the admin. This is
to make this process future-proof and safeguard from programmer's
mistake of using `msg.sender` in initialize function.

This changes the upgrade promise a little. With `onlyAdmin` restriction
we promise to do not upgrade the system sooner than one day after we
announce the upgrade but there is no guarantee this will be carried out
in the timeframe.

Still, we had a similar chance without `onlyAdmin` but it was just more
expensive - we could announce a new upgrade with identical
implementation contract and this would effectively postpone the upgrade.