Adjust stake amount and add stake-capital
Piotr rightly pointed out that even though we were providing minimum
stake to accounts it's only enough for a single ticket submission.
Since we're operating on a limited number of clients they need to be
able to submit more than one ticket.  Here we increase the stake amount
to allow for that.

We also added a new account for stake-capital, who is getting a bundle
today.