Add real dumb account creation script
I didn't want to create the accounts we need for system tests by hand.
Here is a half-assed script that will produce accounts with unique
passphrases.   Account info is written to file if you need it after the
script runs, keyfiles are written in the default extra long geth format.