Expect __beaconCallback(uint256) callback function
Instead of allowing the customer to pass callback signature of their
choice we now expect that the callback contract delcares public
__beaconCallback(uint256) function.

Previously, KeepRandomBeaconServiceImplV1 allowed senders to specify
an arbitrary method and contract that would receive a callback once
the beacon generated a relay entry.

Arbitrary callbacks could be used to force the service contract to
execute many functions within the keep contract system, for example
functions with onlyServiceContract modifier in the operator contract.