Migrate contracts independently of client build
In olden times when client builds were fast and client deployment was
tied in more closely to client builds, contract migration was contingent
on a successful build + test of the Go client. Now the client build and
the contract migration both take about 10 minutes, and deployment of new
clients is gated on the initcontainer build and image publish. As such,
the contract migration now proceeds independently of the client build,
but publishing the package for the contracts is contingent on a
successful build.