Introduce separate build-and-test workflow
This workflow only runs on non-master branches, while the existing
build-test-migrate-publish-keep-dev workflow runs on the master branch.
Notably, the build-test workflow only builds and tests the Go client and
builds the token dashboard app based on the latest-published contract
artifacts.

This takes care of a limitation in build-test-migrate-publish-keep-dev
where there was a need to depend on the NPM publish for the token
dashboard build, but only on master builds. The workflows are now fully
split so that master builds can get the custom behavior.