Fix npm bumping issue with single-match versions
The jq command in use expected the output from npm show to EITHER be
empty (resulting in an empty jq output) OR be a JSON array. It turns
out, if there's only one matching version in npm, npm show returns a
JSON string instead of a JSON array with one string. To handle this, the
JSON is now always wrapping the JSON in an array, and then that array is
flattened to find the latest version.