Add permanently approved operator contract state
This is a draft version for discussion, and still needs tests. It allows
operator contracts to be approved _permanently_ on the registry, making
them ineligible for panic-button-based disabling. To make it clear that
this is a more dangerous contract to approve, the
`authorizeOperatorContract` method on `TokenStaking` is adjusted to only
authorize non-permanently-approved contracts. To authorize a
permanently-approved contract, the
`permanentlyAuthorizeOperatorContract` method is introduced on
`TokenStaking`. Docs reflect the additional dangers of authorizing such
a contract.