TestArrayUtils and TestModUtils moved to the main test directory
There is no poing in keeping them in a separate directory given that
they test two different contracts and we don't have a convention to
reflect the same directory structure in `solidity/test` and
`solidity/contracts`. Also, other files such as `TestAltBn128.js` or
`TestBLS.js` are not kept in separate directories.