Use TokenGrant instead of TokenStaking
Rather than the faucet stake tokens for folks we want them to checkout
our token dashboard.  To help facilitate that we've updated the faucet
function to issue grants.  For a passed account you can get a 300k KEEP
grant, if you're at this grant value then the function will exit, if
you're somewhere in between it will take the difference of your balance
and 300k then grant that amount.