Improve terminated group handling
Rename `terminatedGroups` to `activeTerminatedGroups`
to reflect the fact that expired groups get removed from it.
Pack group termination status in the `Group` struct itself
by reducing blockheight size from 256 to 64 bits.

Add `groupIndices` mapping the group public key to its index
(flagged with the most significant bit to distinguish the index 0)
to avoid O(N) loops when looking for the index of a group
identified by the group pubkey.