Removed on-chain relay entry generation time
We are now submitting relay entry as soon as we are able to compute it.
We need t+1 valid entries and it should not take more than a couple of
seconds to generate entry. We no longer need relay entry generation time
in the contract, as the entry should be generated by clients almost
instantly.