Combine `approveAndCall` and direct interface in ManagedGrantFactory
`approveAndCall` is a single step but cannot return values.
In dApp usage this is not a problem,
as relevant info can be returned in events events and parsed off-chain.
However, creating managed grants on-chain
would be difficult without further workarounds
(such as the usage of `numGrants()` in the factory itself).

Providing both `receiveApproval` and `createGrant`
accommodates both on-chain and dApp use-cases
and ensures that whomever builds another layer of contracts on top of this
doesn't need to figure out workarounds of their own.