Use Circle Context for Docker image tag
We want to make sure Docker images built for versioned environments are
actually versioned themselves.  Here we introduce a new Circle Context
var to help us with that.  Circle builds that run against a git tag generate
a Circle env var CIRCLE_TAG.  This is perfect for our Docker image
tagging needs.  However, the CIRCLE_TAG env var isn't generated for
builds that don't run on a tag.  e.g. all keep-dev / master builds.  We
already use env based Circle contexts to factor out env specific configs
from the Circle config space, so it's a natural fit to handle this in
the same way.  The Circle Context env var is DOCKER_IMAGE_TAG.  For
keep-dev it's set to latest, and for keep-test it's set to $CIRCLE_TAG.