Refactor grant request lifetime, status, + errors
In particular, status codes reflect either request issues (400) or
server issues (500) when appropriate. Additionally, the grant amount is
handled as a BN end-to-end.

Last but not least, the request endpoint returns a link to an Etherscan
transaction immediately rather than waiting for the transaction to be
mined. This helps requests finish quickly rather than depending on
Ropsten mining times.