net.Firewall
So far we had just one rule we had to enforce when establishing a
connection: the remote peer needs to have a minimum stake and the
minimum stake has to be maintained for the entire lifetime of the
connection.

It is still the case but we are discovering that more rules are needed
and that we need exceptions from some of the rules in certain
circumstances.

For example, for ECDSA keep we should require a minimum stake if the
client has no active keeps. If the given client does no longer have a
minimum stake but it has a number of active KEEPs we should let the
client to join the network and execute signing for those remaining
keeps.

To do that, we introduce net.Firewall interface which is a facade behind
which we can hide even the most sophisticated remote peer validation
rules. Instead of passing chain.StakeMonitor down to libp2p internals,
we now pass net.Firewall with a specific minimum stake policy defined in
firewall.go

From the functional side, nothing changed here. However, this will allow
to define more complex connection filtering rules for keep-ecdsa.