Permit assignment of grantManager when creating token grants
ManagedGrantFactory is the creator of the TokenGrants
that get wrapped in a ManagedGrant.
If the token grants automatically assign the token sender as the grant manager,
the ManagedGrantFactory will be performing this role.
This would require duplicating grant management code in the ManagedGrantFactory.

To avoid this, add `_grantManager` as the first argument
in the `_extraData` of the `receiveApproval` call.
The managed grant factory automatically assigns the sender as the manager,
and passes this on to the token grant contract.

Add tests to ensure that not only is the managed grant created correctly,
but the manager and unlocking schedule of the grant are correct too.