Per-contract panic buttons that can be disabled
We now allow to specify a panic button per contract. Such individual
panic buttons can be disabled what makes the operator contract
permanently approved. There is a default panic button value that is
assigned to each new operator contract.

The only role which can disable panic button for operator contract is
governance. We don't want the operator contract upgrader to have a power
of permanently approving operator contracts if this action can not be
later undone.