Added tests from seizing/slashing empty stake
We need to make sure that if slashing/seizing is executed for operator not
having anything on stake, we are not going to fail.
It is possible that two slashings will be executed quickly one after
another. For example, beacon relay entry timeout happens and
immediatelly after that someone reports signature fraud in ECDSA keep.