Let the contract to delegate authority to another contract directly
In some cases, when operator contract creates another contract (e.g.
factory) and wants to pass its authority to the created contract, going
through claim mechanism may be a little too complex and cause problems
with gas estimation. We solve it by another function:
`delegateAuthority` that has the same effect as
`claimDelegatedAuthority` except that it is called by the contract
granting the authorization and not by the contract claiming it.