Broadcast channel relay
Here we introduce a new behavior of the client
which now subscribes to all broadcast channels and
relays received messages to connected peers. This
way the client acts like an auxiliary node even
if they doesn't take part in given protocol
execution. This change should make the network
more robust.