Properly handle the case when the last group gets expired.
We had a bug in expireOldGroups function of Groups library allowing the
expiredGroupOffset to go too far and read from groups array at
non-existing index. It resulted in blocking the beacon forever by
requestRelayEntry reverting every single time it's called.

There are couple of fixes here to fix this problem.

First of all, we fix the loop so that it never reads from groups array
with an index equal or greater to groups array length. Secondly, in
selectGroup function, we first expireOldGroups before evaluation if we
have any active groups to proceed with. Last but not least, in genesis
function we now run group expiration before checking the number of
active groups.

There is still a little chicked and egg problem here addressed with
appropriate documentation on numberOfGroups function. Specifically,
numberOfGroups may return the number > 0 and still, request relay entry
will fail with no active groups message. The reason for that is because
numberOfGroups takes into account the current state in the contract and
requestRelayEntry modifies this state by running group expiration. If
all group expired, requestRelayEntry can't continue and it reverts
without updating chain state. There is no easy way to address this issue
without a solid refactoring so I decided to explain it in numberOfGroups
docs for now. If such situation happens, it's now possible to run
genesis() to fix it as we'll first run group expiration there.