Setup GKE cluster an VPN
Includes a regional GKE cluster and VPN to connect to the cluster.

There are a couple differences between this setup and lesser
environments worth noting:

1. We moved away from the openvpn module.  It's a single resource, it
doesn't need to be provisioned from a module.  the helm_release is
configured directly in the vpn.tf file.

2. We're using updated Helm and Kubernetes providers.  I wanted to give
helm v3 a shot to see if we could keep from poisoning the new env with
all of the Tiller overhead nonsense.  This works, so we're going to keep
it.  We can use what's configured here as a template for migrating
lesser envs.