Allow to specify service contract keeper in the Registry
Some operator contracts, such as keep random beacon operator contract,
specify a list of service contracts it wants to work with. Service
contracts can be added and removed from that list.

Right now, keep random beacon operator contract has an owner which can
add and remove contracts from that list and owner is set to the address
which deployed the contract.

We do not want to have our contracts ownable so we introduce service
contract keeper role, set per the operator contract which will be used
to manage service contract list.